Update README.md
The program is designed to read in an array of floating-point values (wind speeds), and sort the values using bubble sort and selection sort algorithms, calculating the time it took for each algorithm and displaying the number of comparisons and swaps each algorithm took. 

The second part of the program takes in the unsorted array data, I used the data in the data.h header file, and by using an array divided the data into five equally spaced partitions and displays it in a user-friendly format.